Add handling for retriable jobs.
Why
If a job is being requeued, in order to reduce the noise in the failed, you
might not want it to  hit the failed queue until it has exceeded the retry
limit.

The change addresses the need by
This change allows the setting of flags for retry_limit and retry_counts. While
the retry count is below the retry limit, the job failure is silenced. When it
hits the retry limit, then the the job is pushed onto the failed queue.
This change also adds a retry helper method that sets the methods for
retry_limit and retry_count.